+++ .codacy/codacy.yaml
@@ -1,9 +1,9 @@
 runtimes:
-  - java@17.0.10
-  - node@22.2.0
-  - python@3.11.11
+    - java@17.0.10
+    - node@22.2.0
+    - python@3.11.11
 tools:
-  - lizard@1.17.19
-  - pmd@6.55.0
-  - semgrep@1.78.0
-  - trivy@0.59.1
+    - lizard@1.17.19
+    - pmd@6.55.0
+    - semgrep@1.78.0
+    - trivy@0.59.1
